change type of BlockNumber from 'u64' to 'u32'

+++ as-packages/lang/assembly/env/BlockNumber.ts
@@ -3,6 +3,5 @@
  * @author liangqin.fan@gmail.com
  */
 
-import { UInt64 } from "as-scale-codec";
 
-export type BlockNumber = UInt64;
+export type BlockNumber = u32;
